* cvs merge * revconnect merge

+++ trunk/client/Encoder.cpp
@@ -33,7 +33,7 @@ string& Encoder::toBase32(const u_int8_t* src, size_t len, string& dst) {
 	for(i = 0, index = 0; i < len;) {
 		/* Is the current word going to span a byte boundary? */
 		if (index > 3) {
-			word = (src[i] & (0xFF >> index));
+			word = (u_int8_t)(src[i] & (0xFF >> index));
 			index = (index + 5) % 8;
 			word <<= index;
 			if ((i + 1) < len)
@@ -41,7 +41,7 @@ string& Encoder::toBase32(const u_int8_t* src, size_t len, string& dst) {
 
 			i++;
 		} else {
-			word = (src[i] >> (8 - (index + 5))) & 0x1F;
+			word = (u_int8_t)(src[i] >> (8 - (index + 5))) & 0x1F;
 			index = (index + 5) % 8;
 			if (index == 0)
 				i++;
@@ -84,6 +84,7 @@ void Encoder::fromBase32(const char* src, u_int8_t* dst, size_t len) {
 		}
 	}
 }
+
 
 bool Encoder::isBase32(const char* src)
 {
